Updated simpleui.py w/ new imports
I added the maya.OpenMayaUI import and I also added wrapInstance from shiboken2. With these I added the function of Maya_Main_Window, which helps me insure that the UI I make is a child of the Maya window. That way Maya won't hide the UI when I'm using Maya.