Updated smartsave.py w/ new method & filename details
I added the text boxes for creating a filename for the smart save. They are laid out in the header UI from previously using the grid layout format. I added a spinbox for the version number. All of the header creating code was also moved to a new method called _create_filename_headers.